chore: #277 - DependencyValues 파일 분리
